backend/requirements.txt
fastapi
uvicorn
asyncpg
httpx
python-dotenv